# specs/openapi.yaml
openapi: 3.0.3
info:
  title: 日程管理系统 API
  version: 3.5.0
  description: |
    v3.3 接口定义 — 前后端契约唯一真相源。

    **v3.3 变更**：
    - 新增任务看板 API：`/tasks`、`/tasks/{id}`、`/tasks/{id}/move`
    - 新增 4 个 schema：TaskProfile、CreateTaskRequest、UpdateTaskRequest、MoveTaskRequest

    **v3.2 变更**：
    - 新增宠物养成 API：`/pets/me`、`/pets/me/interact`、`/shop/items`、`/shop/purchase`
    - 新增 4 个 schema：PetProfile、InteractionResult、ShopItem、PurchaseResult

    **v3.1 变更**：
    - Event 新增 `status` 字段（PLANNED / COMPLETED / CANCELLED），支持"标记完成"。
    - `GET /events` 新增 `tagId`、`status` 过滤参数。
    - 已完成/已取消的日程不再触发提醒，也不参与时间冲突检测。

    **响应约定**：
    - 成功响应直接返回业务数据：单条返回对象，列表返回数组。
    - HTTP 状态码承载语义（200 / 201 / 204 / 400 / 401 / 403 / 404 / 409 / 500）。
    - 异常响应统一为 `ApiResponse`：`{ code, message }`。

    **认证（v3.0）**：
    - 除 `/auth/register` / `/auth/login` / `/auth/refresh` / `/sse/notifications` 外，
      所有端点要求 `Authorization: Bearer <accessToken>`。
    - `/sse/notifications` 不走 Bearer（EventSource 不支持自定义 header），
      改用登录时下发的 `dsa_sse_session` HttpOnly Cookie 鉴权。
    - 用户数据相互隔离：所有 event / category / tag 操作以当前用户为范围。

    变更历史见 `specs/CHANGELOG.md`。
servers:
  - url: /api/v1
security:
  - bearerAuth: []
paths:
...
  /tags:
    get:
      operationId: listTags
      tags: [Tag]
      summary: 获取当前用户的所有标签
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TagResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      operationId: createTag
      tags: [Tag]
      summary: 创建标签
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TagCreateRequest'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TagResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
  /tags/{id}:
    put:
      operationId: updateTag
      tags: [Tag]
      summary: 更新标签
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TagCreateRequest'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TagResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
    delete:
      operationId: deleteTag
      tags: [Tag]
      summary: 删除标签
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: 删除成功
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /pets/me:
    post:
      operationId: createPet
      tags: [Pet]
      summary: 创建当前用户的宠物
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePetRequest'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PetProfile'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
    get:
      operationId: getMyPet
      tags: [Pet]
      summary: 获取当前用户的宠物详情
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PetProfile'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      operationId: updatePet
      tags: [Pet]
      summary: 更新宠物（仅支持改名）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePetRequest'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PetProfile'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /pets/me/interact:
    post:
      operationId: interactWithPet
      tags: [Pet]
      summary: 与宠物互动（喂食/玩耍）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InteractRequest'
      responses:
        '200':
          description: 互动成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InteractionResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /pets/me/accessory:
    delete:
      operationId: unequipAccessory
      tags: [Pet]
      summary: 取下当前配饰（幂等，未装备时同样成功）
      responses:
        '204':
          description: 取下成功
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /pets/me/rewards:
    post:
      operationId: grantPetReward
      tags: [Pet]
      summary: 领取行为奖励（幂等，granted=false 表示未发放）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GrantRewardRequest'
      responses:
        '200':
          description: 发放结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RewardResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /shop/purchase:
    post:
      operationId: purchaseItem
      tags: [Shop]
      summary: 购买物品（即时消费模式）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PurchaseRequest'
      responses:
        '200':
          description: 购买成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PurchaseResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /tasks:
    get:
      operationId: listTasks
      tags: [Task]
      summary: 查询任务列表
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [TODO, IN_PROGRESS, DONE]
          description: 按状态过滤
        - name: priority
          in: query
          required: false
          schema:
            type: string
            enum: [LOW, MEDIUM, HIGH, URGENT]
          description: 按优先级过滤
        - name: tagId
          in: query
          required: false
          schema:
            type: integer
            format: int64
          description: 按标签过滤
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TaskProfile'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      operationId: createTask
      tags: [Task]
      summary: 创建任务
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskRequest'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskProfile'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /tasks/{id}:
    put:
      operationId: updateTask
      tags: [Task]
      summary: 更新任务
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTaskRequest'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskProfile'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      operationId: deleteTask
      tags: [Task]
      summary: 删除任务
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: 删除成功
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /tasks/{id}/move:
    patch:
      operationId: moveTask
      tags: [Task]
      summary: 移动任务到目标列
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MoveTaskRequest'
      responses:
        '200':
          description: 移动成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskProfile'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        登录后将 `LoginResponse.accessToken` 放入 `Authorization: Bearer <token>` 头。
        access token 默认有效期 15 分钟；过期后用 refresh token 调 `/auth/refresh` 续签。
  responses:
    BadRequest:
      description: 请求参数不合法
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
    Unauthorized:
      description: 未认证或 token 失效
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
    NotFound:
      description: 资源不存在或无权访问
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
    Conflict:
      description: 资源冲突（用户名/邮箱/分类名/标签名重复）
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
  schemas:
...
    TagCreateRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          maxLength: 30
        color:
          type: string
    TagResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        color:
          type: string
...
    CreatePetRequest:
      type: object
      required: [species, name]
      properties:
        species:
          $ref: '#/components/schemas/PetSpecies'
        name:
          type: string
          maxLength: 30
    UpdatePetRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          maxLength: 30
    PetSpecies:
      type: string
      enum: [ORANGE_CAT, SHIBA_INU]
    PetProfile:
      type: object
      properties:
        id:
          type: integer
          format: int64
        species:
          type: string
        name:
          type: string
        experience:
          type: integer
        level:
          type: integer
        mood:
          type: integer
        hunger:
          type: integer
        coins:
          type: integer
        currentAccessory:
          type: integer
          format: int64
        lastInteractedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
    InteractRequest:
      type: object
      required: [type]
      properties:
        type:
          type: string
          enum: [FEED, PLAY]
        itemId:
          type: integer
          format: int64
          description: 喂食时指定的食物 ID（可选，默认最便宜食物）
        quantity:
          type: integer
          default: 1
          minimum: 1
          maximum: 10
    InteractionResult:
      type: object
      properties:
        moodChange:
          type: integer
        hungerChange:
          type: integer
        experienceGain:
          type: integer
        coinChange:
          type: integer
        newMood:
          type: integer
        newHunger:
          type: integer
        newExperience:
          type: integer
        newCoins:
          type: integer
...
    PurchaseRequest:
      type: object
      required: [itemId]
      properties:
        itemId:
          type: integer
          format: int64
        quantity:
          type: integer
          default: 1
          minimum: 1
    PurchaseResult:
      type: object
      properties:
        success:
          type: boolean
        itemName:
          type: string
        quantity:
          type: integer
        totalCost:
          type: integer
        newCoins:
          type: integer
        newMood:
          type: integer
        newHunger:
          type: integer
        newExperience:
          type: integer
        equippedAccessoryId:
          type: integer
          format: int64
          description: 装备购买时回传所装备配饰 id（FOOD 购买为 null）
    GrantRewardRequest:
      type: object
      required: [source, refId]
      properties:
        source:
          type: string
          enum: [TASK_COMPLETED, EVENT_COMPLETED, EVENT_CANCELLED, FOCUS_COMPLETED, DAILY_CHECKIN, HABIT_CHECKED]
        refId:
          type: string
          maxLength: 64
          description: 业务引用标识（幂等键），如 taskId / eventId / habitId / 日期
    RewardResult:
      type: object
      required: [granted, coinChange, experienceGain, moodChange, newCoins, newExperience, newMood]
      properties:
        granted:
          type: boolean
          description: false 表示未发放（无宠物或幂等键已命中）
        coinChange:
          type: integer
        experienceGain:
          type: integer
        moodChange:
          type: integer
        newCoins:
          type: integer
        newExperience:
          type: integer
        newMood:
          type: integer
    TaskProfile:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        description:
          type: string
        status:
          type: string
        priority:
          type: string
        sortOrder:
          type: integer
        dueDate:
          type: string
          format: date
        tags:
          type: array
          items:
            $ref: '#/components/schemas/TagResponse'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateTaskRequest:
      type: object
      required: [title]
      properties:
        title:
          type: string
          maxLength: 200
        description:
          type: string
        priority:
          type: string
          enum: [LOW, MEDIUM, HIGH, URGENT]
        dueDate:
          type: string
          format: date
        tagIds:
          type: array
          items:
            type: integer
            format: int64
    UpdateTaskRequest:
      type: object
      properties:
        title:
          type: string
          maxLength: 200
        description:
          type: string
        priority:
          type: string
          enum: [LOW, MEDIUM, HIGH, URGENT]
        dueDate:
          type: string
          format: date
        tagIds:
          type: array
          items:
            type: integer
            format: int64
    MoveTaskRequest:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [TODO, IN_PROGRESS, DONE]
        sortOrder:
          type: integer
    ApiResponse:
      type: object
      description: 错误响应包装
      properties:
        code:
          type: integer
        message:
          type: string